+++ noodles-bcf/examples/bcf_query.rs
@@ -6,13 +6,11 @@
 
 use std::{
     env,
-    fs::File,
     io::{self, BufWriter},
     path::PathBuf,
 };
 
 use noodles_bcf as bcf;
-use noodles_csi as csi;
 use noodles_vcf as vcf;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -21,13 +19,11 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = args.nth(1).map(PathBuf::from).expect("missing src");
     let raw_region = args.next().expect("missing region");
 
-    let mut reader = File::open(&src).map(bcf::Reader::new)?;
+    let mut reader = bcf::indexed_reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?;
 
-    let index = csi::read(src.with_extension("bcf.csi"))?;
-
     let region = raw_region.parse()?;
-    let query = reader.query(&header, &index, &region)?;
+    let query = reader.query(&header, &region)?;
 
     let stdout = io::stdout().lock();
     let mut writer = vcf::Writer::new(BufWriter::new(stdout));