+++ noodles-bcf/src/indexed_reader.rs
@@ -1,3 +1,9 @@
+//! Indxed BCF reader.
+
+mod builder;
+
+pub use self::builder::Builder;
+
 use std::io::{self, Read, Seek};
 
 use noodles_bgzf as bgzf;